requirements.txt
dj-database-url==0.5.0
Django==2.1.1
django-debug-toolbar==1.10.1
django-dotenv==1.4.2
django-sslserver==0.20
djangorestframework==3.8.2
gunicorn==19.9.0
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
sqlparse==0.2.4
whitenoise==4.0
zbarlight==2.1
